fix: guard against zero dimensions, stack overflow, and PIT wrap
- `compute_letterbox`: return zero layout when any dimension is zero to avoid division by zero.
- `hda_feed_silence`: replace stack-allocated 16 KiB buffer with a static one to prevent stack overflow on small kernel stacks.
- PIT calibration: measure 20 ms of PIT ticks instead of waiting for a full 55 ms counter wrap to avoid hangs if preempted.
- compositor `alpha_blend!` macro: remove internal `continue`, return a boolean indicating whether further blending is required so callers can control loop flow.
- wallpaper shell command: print error message when a preset is not found.

+++ solvent/src/lib.rs
@@ -230,7 +230,7 @@ pub struct RuntimeState {
 
     /// Lightweight cursor save buffer (16×16 pixels) for overlay mode
     /// cursor movement without full re‑renders.
-    pub cursor_save_buf: [u32; lattice::cursor::Cursor::SIZE * lattice::cursor::Cursor::SIZE],
+    pub cursor_save_buf: [u32; lattice::cursor::Cursor::SIZE as usize * lattice::cursor::Cursor::SIZE as usize],
     pub cursor_save_x: i32,
     pub cursor_save_y: i32,
     pub cursor_save_valid: bool,
